Recuperar clientes de quotes

+++ src/main/java/com/nidara/sabanasblancas/gestion/model/dtos/PagedResult.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class PagedResult<T> {
 
-    public static final long DEFAULT_PAGE = 0;
+    public static final int DEFAULT_PAGE = 0;
 
     public static final int DEFAULT_SIZE = 20;
